update(scripts): git/after.sh and karabiner install script

+++ .config/git/after.sh
@@ -41,6 +41,10 @@ function git_config() {
 git config --global include.path "$HOME/.config/git/config_shared"
 git_config
 
+if [[ -e "$HOME/.git-templates/git-secrets" ]]; then
+  exit 0
+fi
+
 if command -v git-secrets > /dev/null 2>&1; then
   git secrets --install ~/.git-templates/git-secrets
   git config --global init.templatedir "$HOME/.git-templates/git-secrets"
